Split log events and beforeExit
