+++ packages/client/src/runtime/core/engines/library/LibraryEngine.ts
@@ -342,14 +342,10 @@ You may have to run ${green('prisma generate')} for your changes to take effect.
     return str
   }
 
-  on<E extends EngineEventType>(event: E, listener: EngineEventCallback<E>): void {
-    if (event === 'beforeExit') {
-      throw new Error(
-        '"beforeExit" hook is not applicable to the library engine since Prisma 5.0.0, it is only relevant and implemented for the binary engine. Please add your event listener to the `process` object directly instead.',
-      )
-    } else {
-      this.logEmitter.on(event, listener)
-    }
+  override onBeforeExit() {
+    throw new Error(
+      '"beforeExit" hook is not applicable to the library engine since Prisma 5.0.0, it is only relevant and implemented for the binary engine. Please add your event listener to the `process` object directly instead.',
+    )
   }
 
   async start(): Promise<void> {
